Stable ground of pyeleven
The build off these images has been broken since 25th of March 2024. (And will continute to fail until #2 is merged).

In order to not overwrite the old images which are built with python2 I think it's wise to start using the tagged version of pyeleven. Probably wise to use a tagged version in production anyway.